requirements.txt
asgiref==3.4.1
Django==4.0
keras==2.7.0
numpy==1.21.4
Pillow==8.4.0
sqlparse==0.4.2
tzdata==2021.5
